Refactor: decomposed ObjectAttributes component (#608)
